project/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
bcrypt==4.1.2
pyjwt==2.8.0
PyPDF2==3.0.1
python-docx==1.1.2
SpeechRecognition==3.10.4
